requirements.txt
python-telegram-bot>=12.8
pydrive>=1.3.1
pygsheets>=2.0.3.1
pandas>=0.25.3
tabulate~=0.8.7
numpy~=1.19.4
